django-backend/pyproject.toml
[project]
name = "django-backend"
version = "0.1.0"
description = "Django backend for Collect Map API"
requires-python = ">=3.12"
dependencies = [
    "django>=5.0.0",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.3.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.2",
    "pillow>=11.3.0",
    "numpy>=2.3.3",
    "requests>=2.32.3",
    "inflection>=0.5.1",
    "uritemplate>=4.1.1",
]
